:art: Split helper into separate file
