+++ packages/loader-js/src/helpers.js
@@ -1,6 +1,7 @@
 /* @flow */
 
 import * as t from 'babel-types'
+import type { File, FileChunk } from 'pundle-api/types'
 
 export function getName(obj: Object): string {
   if (typeof obj.name === 'string') {
@@ -18,6 +19,7 @@ export function getName(obj: Object): string {
 
 const STRING_REGEX = /^"[^"]*"$/
 
+// TODO: use babel-template
 export function getParsedReplacement(rawValue: any): Object {
   let parsedValue
   if (STRING_REGEX.test(rawValue)) {
@@ -30,4 +32,33 @@ export function getParsedReplacement(rawValue: any): Object {
     parsedValue = t.identifier(rawValue)
   }
   return parsedValue
+}
+
+export function processSplit(file: File, chunks: Array<FileChunk>, path: Object) {
+  const [nodeEntry, nodeCallback, nodeName] = path.node.arguments
+
+  const chunk = {
+    name: nodeName ? nodeName.value : this.getNextUniqueID().toString(),
+    entry: [],
+    imports: [],
+  }
+  nodeEntry.elements.forEach(element => {
+    const request = this.getImportRequest(element.value, file.filePath)
+    chunk.entry.push(request)
+    element.value = request.id.toString()
+  })
+  if (nodeCallback && nodeCallback.params.length) {
+    const nodeCallbackParam = nodeCallback.params[0]
+    path.scope.traverse(nodeCallback, {
+      CallExpression: ({ node, scope }) => {
+        if (getName(node.callee) === nodeCallbackParam.name && !scope.getBinding(nodeCallbackParam.name)) {
+          const request = this.getImportRequest(node.arguments[0].value, file.filePath)
+          chunk.imports.push(request)
+          node.arguments[0].value = request.id.toString()
+        }
+      },
+    })
+  }
+
+  chunks.push(chunk)
 }
